+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-07-02 21:48:49
-// Commit: test(i18n): implement widget integration tests for i18n
+// Updated at: 2026-07-03 09:20:16
+// Commit: chore(i18n): update dependencies and cleanup unused imports in i18n
 
 class i18nComponentState {
-  final String commitRef = '2497';
-  final String lastMessage = 'test(i18n): implement widget integration tests for i18n';
-  final DateTime timestamp = DateTime.parse('2026-07-02T21:48:49+05:00');
+  final String commitRef = '2503';
+  final String lastMessage = 'chore(i18n): update dependencies and cleanup unused imports in i18n';
+  final DateTime timestamp = DateTime.parse('2026-07-03T09:20:16+05:00');
 }